requirements.txt
python-dotenv~=0.17.1
numpy~=1.20.2